requirements.txt
discord.py
python-dotenv
requests
google-generativeai
PyNaCl
yt-dlp
isolate
catholic_mass_readings
lavalink